// Boardlight/src/api/generated/models/shopIndexOwnedItemsResponse.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */
import type { ShopIndexOwnedItemsResponseProduct } from './shopIndexOwnedItemsResponseProduct';

export interface ShopIndexOwnedItemsResponse {
  id?: number;
  userId?: string;
  product?: ShopIndexOwnedItemsResponseProduct;
  usedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}